+++ tests/src/Opulence/Bootstrappers/BootstrapperRegistryTest.php
@@ -8,7 +8,7 @@
  */
 namespace Opulence\Bootstrappers;
 
-use Opulence\Applications\Environments\Environment;
+use Opulence\Environments\Environment;
 use Opulence\Tests\Bootstrappers\Mocks\EagerBootstrapper;
 use Opulence\Tests\Bootstrappers\Mocks\EagerFooInterface;
 use Opulence\Tests\Bootstrappers\Mocks\LazyBootstrapper;