+++ tests/src/Opulence/Framework/Testing/PhpUnit/Console/Mocks/ApplicationTestCase.php
@@ -9,13 +9,13 @@
 namespace Opulence\Tests\Framework\Testing\PhpUnit\Console\Mocks;
 
 use Opulence\Applications\Application;
-use Opulence\Applications\Environments\Environment;
 use Opulence\Applications\Tasks\Dispatchers\Dispatcher as TaskDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
 use Opulence\Bootstrappers\BootstrapperRegistry;
 use Opulence\Bootstrappers\Caching\ICache as BootstrapperCache;
 use Opulence\Bootstrappers\Dispatchers\Dispatcher;
 use Opulence\Bootstrappers\Paths;
+use Opulence\Environments\Environment;
 use Opulence\Framework\Bootstrappers\Console\Commands\CommandsBootstrapper;
 use Opulence\Framework\Bootstrappers\Console\Composer\ComposerBootstrapper;
 use Opulence\Framework\Testing\PhpUnit\Console\ApplicationTestCase as BaseApplicationTestCase;
@@ -46,19 +46,19 @@ class ApplicationTestCase extends BaseApplicationTestCase
         ]);
         $taskDispatcher = new TaskDispatcher();
         // Purposely set this to a weird value so we can test that it gets overwritten with the "test" environment
-        $environment = new Environment("foo");
+        $this->environment = new Environment("foo");
         $this->container = new Container();
         $this->container->bind(Paths::class, $paths);
         $this->container->bind(TaskDispatcher::class, $taskDispatcher);
-        $this->container->bind(Environment::class, $environment);
+        $this->container->bind(Environment::class, $this->environment);
         $this->container->bind(BootstrapperCache::class, $this->getMock(BootstrapperCache::class));
         $this->container->bind(RouteCache::class, $this->getMock(RouteCache::class));
         $this->container->bind(ViewCache::class, $this->getMock(ViewCache::class));
         $this->container->bind(IContainer::class, $this->container);
-        $this->application = new Application($taskDispatcher, $environment);
+        $this->application = new Application($taskDispatcher);
 
         // Setup the bootstrappers
-        $bootstrapperRegistry = new BootstrapperRegistry($paths, $environment);
+        $bootstrapperRegistry = new BootstrapperRegistry($paths, $this->environment);
         $bootstrapperDispatcher = new Dispatcher($taskDispatcher, $this->container);
         $bootstrapperRegistry->registerEagerBootstrapper(self::$bootstrappers);
         $taskDispatcher->registerTask(