+++ tests/src/Opulence/Framework/Testing/PhpUnit/Http/Mocks/ApplicationTestCase.php
@@ -9,13 +9,13 @@
 namespace Opulence\Tests\Framework\Testing\PhpUnit\Http\Mocks;
 
 use Opulence\Applications\Application;
-use Opulence\Applications\Environments\Environment;
 use Opulence\Applications\Tasks\Dispatchers\Dispatcher as TaskDispatcher;
 use Opulence\Applications\Tasks\TaskTypes;
 use Opulence\Bootstrappers\BootstrapperRegistry;
 use Opulence\Bootstrappers\Dispatchers\Dispatcher;
 use Opulence\Bootstrappers\Paths;
 use Opulence\Debug\Exceptions\Handlers\ExceptionHandler;
+use Opulence\Environments\Environment;
 use Opulence\Framework\Bootstrappers\Http\Requests\RequestBootstrapper;
 use Opulence\Framework\Bootstrappers\Http\Routing\RouterBootstrapper;
 use Opulence\Framework\Debug\Exceptions\Handlers\Http\IExceptionRenderer;
@@ -46,16 +46,16 @@ class ApplicationTestCase extends BaseApplicationTestCase
         ]);
         $taskDispatcher = new TaskDispatcher();
         // Purposely set this to a weird value so we can test that it gets overwritten with the "test" environment
-        $environment = new Environment("foo");
+        $this->environment = new Environment("foo");
         $this->container = new Container();
         $this->container->bind(Paths::class, $paths);
         $this->container->bind(TaskDispatcher::class, $taskDispatcher);
-        $this->container->bind(Environment::class, $environment);
+        $this->container->bind(Environment::class, $this->environment);
         $this->container->bind(IContainer::class, $this->container);
-        $this->application = new Application($taskDispatcher, $environment);
+        $this->application = new Application($taskDispatcher);
 
         // Setup the bootstrappers
-        $bootstrapperRegistry = new BootstrapperRegistry($paths, $environment);
+        $bootstrapperRegistry = new BootstrapperRegistry($paths, $this->environment);
         $bootstrapperDispatcher = new Dispatcher($taskDispatcher, $this->container);
         $bootstrapperRegistry->registerEagerBootstrapper(self::$bootstrappers);
         $taskDispatcher->registerTask(